requirements.txt
# Django Modules
django
django-celery-results
django-cors-headers
django-debug-toolbar
django-environ
django-ipware
django-model-utils

# Django Rest Modules
djangorestframework
dj-rest-auth

# Other Modules
celery
coreapi
cryptography
gunicorn
money
requests
passlib
psycopg2
pyopenssl
pytz
validators